core/accounts/api/v1/urls/profiles.py

from django.urls import path
from .. import views
from rest_framework.routers import DefaultRouter
from rest_framework.authtoken.views import ObtainAuthToken
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)


urlpatterns = [
    path("profile/user", views.ProfileApiView.as_view(), name="profile")
]
